vim-syntax: highlight trailing : on TODO, MARK, FIXME
These three commands have semantic meaning for swift parsers, for
example, they're used to show structure in Xcode. But, they only have
this meaning when followed by `:`. So, only highlight them when followed
by a colon. This commit also sets `syn iskeyword` to all alphanumeric
characters and : in order to do this.